+++ src/main/java/codesquad/application/config/DatabaseConfig.java
@@ -2,6 +2,7 @@ package codesquad.application.config;
 
 import codesquad.application.bean.Bean;
 import codesquad.application.database.ArticleDatabase;
+import codesquad.application.database.ArticleMemoryDatabase;
 import codesquad.application.database.UserDatabase;
 import codesquad.application.database.UserMemoryDatabase;
 import codesquad.application.database.SessionStorage;
@@ -20,6 +21,6 @@ public class DatabaseConfig {
 
     @Bean
     public ArticleDatabase articleDatabase() {
-        return new ArticleDatabase();
+        return new ArticleMemoryDatabase();
     }
 }
